canbus: fix compile warning.

+++ modules/canbus/vehicle/gem/protocol/global_cmd_69.cc
@@ -61,7 +61,7 @@ Globalcmd69* Globalcmd69::set_pacmod_enable(
 // 'physical_unit': ''}
 void Globalcmd69::set_p_pacmod_enable(
     uint8_t* data, Global_cmd_69::Pacmod_enableType pacmod_enable) {
-  int x = pacmod_enable;
+  uint8_t x = pacmod_enable;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 0, 1);
@@ -80,7 +80,7 @@ Globalcmd69* Globalcmd69::set_clear_override(
 // 'type': 'enum', 'order': 'motorola', 'physical_unit': ''}
 void Globalcmd69::set_p_clear_override(
     uint8_t* data, Global_cmd_69::Clear_overrideType clear_override) {
-  int x = clear_override;
+  uint8_t x = clear_override;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 1, 1);
@@ -99,7 +99,7 @@ Globalcmd69* Globalcmd69::set_ignore_override(
 // 'type': 'enum', 'order': 'motorola', 'physical_unit': ''}
 void Globalcmd69::set_p_ignore_override(
     uint8_t* data, Global_cmd_69::Ignore_overrideType ignore_override) {
-  int x = ignore_override;
+  uint8_t x = ignore_override;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 2, 1);
